# swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "description": "API for easy study new things from school",
        "version": "0.3.4",
        "title": "Learn and test knowledge - API"
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "paths": {
        "/quiz/users/register": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Registration in system",
                "requestBody": {
                    "description": "User data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "userEmail": {
                                        "type": "string"
                                    },
                                    "userPassword": {
                                        "type": "string"
                                    },
                                    "confirmPassword": {
                                        "type": "string"
                                    },
                                    "userGender": {
                                        "type": "string"
                                    },
                                    "userVerification": {
                                        "type": "boolean"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully registered"
                    },
                    "400": {
                        "description": "Error about entered data"
                    },
                    "500": {
                        "description": "System error - user role doesn't exist"
                    }
                }
            }
        },
        "/quiz/users/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Sign In",
                "requestBody": {
                    "description": "User data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "userEmail": {
                                        "type": "string"
                                    },
                                    "userPassword": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "System will return token"
                    },
                    "400": {
                        "description": "Error about entered data"
                    },
                    "404": {
                        "description": "Data not found"
                    }
                }
            }
        },
        "/quiz/users/email": {
            "patch": {
                "tags": [
                    "Users"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "User can change e-mail",
                "requestBody": {
                    "description": "User data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "newUserEmail": {
                                        "type": "string"
                                    },
                                    "userPassword": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data changed successfully"
                    },
                    "400": {
                        "description": "Error about entered data"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Data not found"
                    }
                }
            }
        },
        "/quiz/users/password": {
            "patch": {
                "tags": [
                    "Users"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "User can change password",
                "requestBody": {
                    "description": "User data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "newUserPassword": {
                                        "type": "string"
                                    },
                                    "confirmNewUserPassword": {
                                        "type": "string"
                                    },
                                    "userPassword": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data changed successfully"
                    },
                    "400": {
                        "description": "Error about entered data"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Data not found"
                    }
                }
            }
        },
        "/quiz/users/delete": {
            "patch": {
                "tags": [
                    "Users"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "User can delete account",
                "requestBody": {
                    "description": "User data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "userPassword": {
                                        "type": "string"
                                    },
                                    "confirmPassword": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "400": {
                        "description": "Error about entered data"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Data not found"
                    }
                }
            }
        },
...
    }
}